+++ R/ard_stats_mcnemar_test.R
@@ -2,6 +2,9 @@
 #'
 #' @description
 #' Analysis results data for McNemar's statistical test.
+#' We have two functions depending on the structure of the data.
+#' - `ard_stats_mcnemar_test()` is the structure expected by [`stats::mcnemar.test()`]
+#' - `ard_stats_mcnemar_test_long()` is one row per ID per group
 #'
 #' @param data (`data.frame`)\cr
 #'   a data frame. See below for details.
@@ -11,9 +14,11 @@
 #'   column names to be compared. Independent tests will
 #'   be computed for each variable.
 #' @param ... arguments passed to `stats::mcnemar.test(...)`
+#' @param id ([`tidy-select`][dplyr::dplyr_tidy_select])\cr
+#'   column name of the subject or participant ID
 #'
 #' @return ARD data frame
-#' @export
+#' @name ard_stats_mcnemar_test
 #'
 #' @details
 #' For the `ard_stats_mcnemar_test()` function, the data is expected to be one row per subject.
@@ -23,6 +28,21 @@
 #' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "broom", reference_pkg = "cardx"))
 #' cards::ADSL |>
 #'   ard_stats_mcnemar_test(by = "SEX", variables = "EFFFL")
+#'
+#' set.seed(1234)
+#' cards::ADSL[c("USUBJID", "TRT01P")] |>
+#'   dplyr::mutate(TYPE = "PLANNED") |>
+#'   dplyr::rename(TRT01 = TRT01P) %>%
+#'   dplyr::bind_rows(dplyr::mutate(., TYPE = "ACTUAL", TRT01 = sample(TRT01))) |>
+#'   ard_stats_mcnemar_test_long(
+#'     by = TYPE,
+#'     variable = TRT01,
+#'     id = USUBJID
+#'   )
+NULL
+
+#' @rdname ard_stats_mcnemar_test
+#' @export
 ard_stats_mcnemar_test <- function(data, by, variables, ...) {
   set_cli_abort_call()
 
@@ -59,6 +79,53 @@ ard_stats_mcnemar_test <- function(data, by, variables, ...) {
     }
   ) |>
     dplyr::bind_rows()
+}
+
+#' @rdname ard_stats_mcnemar_test
+#' @export
+ard_stats_mcnemar_test_long <- function(data, by, variables, id, ...) {
+  set_cli_abort_call()
+
+  # check installed packages ---------------------------------------------------
+  check_pkg_installed("broom", reference_pkg = "cardx")
+
+  # check/process inputs -------------------------------------------------------
+  check_not_missing(data)
+  check_not_missing(variables)
+  check_not_missing(by)
+  check_not_missing(id)
+  check_data_frame(data)
+  data <- dplyr::ungroup(data)
+  cards::process_selectors(data, by = {{ by }}, variables = {{ variables }}, id = {{ id }})
+  check_scalar(by)
+  check_scalar(id)
+
+  # if no variables selected, return empty tibble ------------------------------
+  if (is_empty(variables)) {
+    return(dplyr::tibble())
+  }
+  # build ARD ------------------------------------------------------------------
+  lapply(
+    variables,
+    function(variable) {
+      .format_mcnemartest_results(
+        by = by,
+        variable = variable,
+        lst_tidy =
+          cards::eval_capture_conditions({
+            # adding this reshape inside the eval, so if there is an error it's captured in the ARD object
+            data_wide <- .paired_data_pivot_wider(data, by = by, variable = variable, id = id)
+            # performing McNemars test
+            stats::mcnemar.test(x = data_wide[["by1"]], y = data_wide[["by2"]], ...) |>
+              broom::tidy()
+          }),
+        ...
+      )
+    }
+  ) |>
+    dplyr::bind_rows()
+
+
 }
 
 #' Convert McNemar's test to ARD